+++ pakyow-core/CHANGELOG.md
@@ -1,5 +1,10 @@
 # v1.1.0 (unreleased)
 
+  * `chg` **Ignore changes to directories in the filewatcher.**
+
+    *Related links:*
+    - [Pull Request #507][pr-507]
+
   * `chg` **Improve support for ignoring directories in filewatcher, add support for ignoring regular expressions.**
 
     *Related links:*
@@ -655,6 +660,7 @@
     *Related links:*
     - [Pull Request #338][pr-338]
 
+[pr-507]: https://github.com/pakyow/pakyow/pull/507
 [pr-506]: https://github.com/pakyow/pakyow/pull/506
 [pr-505]: https://github.com/pakyow/pakyow/pull/505
 [pr-504]: https://github.com/pakyow/pakyow/pull/504